segfault part 1

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jbelena <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/06 22:27:40 by jbelena           #+#    #+#             */
-/*   Updated: 2019/11/11 17:44:14 by jbelena          ###   ########.fr       */
+/*   Updated: 2019/11/22 16:58:12 by jbelena          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "auxiliar.h"
+#include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	while (s[i])
